create get_tensor_network_state method
Moved methods that absorb weights into TensorNetwork class.

Created `TensorNetwork .get_tensor_network_state()` method
